src/main/java/ua/com/finalproject/controller/EmailController.java
package ua.com.finalproject.controller;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import ua.com.finalproject.service.EmailService;

@RestController
@RequiredArgsConstructor
@RequestMapping("/api/v1/email")
@Slf4j
public class EmailController {
    private final EmailService emailService;

    @PostMapping("/send")
    public String sendEmail(@RequestParam String to, @RequestParam String subject, @RequestParam String text) {
        log.info("Sending email to: {}, with subject: {}", to, subject);
        emailService.sendEmail(to, subject, text);
        log.info("Email successfully sent to: {}", to);
        return "Email successfully sent!";
    }
}
